chore: refactor, so mocking project and application works the same way
