+++ packages/jetbrains-plugin/src/test/kotlin/com/mongodb/jbplugin/ActivatePluginPostStartupActivityTest.kt
@@ -1,7 +1,9 @@
 package com.mongodb.jbplugin
 
+import com.intellij.openapi.project.Project
+import com.mongodb.jbplugin.fixtures.IntegrationTest
 import com.mongodb.jbplugin.fixtures.eventually
-import com.mongodb.jbplugin.fixtures.mockProject
+import com.mongodb.jbplugin.fixtures.withMockedService
 import com.mongodb.jbplugin.observability.probe.PluginActivatedProbe
 import org.junit.jupiter.api.Test
 import org.mockito.Mockito.mock
@@ -11,11 +13,13 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
 
+@IntegrationTest
 class ActivatePluginPostStartupActivityTest {
     @Test
-    fun `emits a plugin activated probe`() = runBlocking {
+    fun `emits a plugin activated probe`(project: Project) = runBlocking {
         val pluginActivatedProbe = mock<PluginActivatedProbe>()
-        val project = mockProject(pluginActivatedProbe = pluginActivatedProbe)
+        project.withMockedService(pluginActivatedProbe)
+
         val listener = ActivatePluginPostStartupActivity(CoroutineScope(Dispatchers.Default))
 
         listener.runActivity(project)
